+++ alertClientDbus/lib/globalData.py
@@ -20,7 +20,7 @@ class GlobalData:
 		self.version = 0.500
 
 		# revision of the used client
-		self.rev = 0
+		self.rev = 1
 
 		# name of this client
 		self.name = "alertR Alert Client Dbus"
@@ -30,7 +30,7 @@ class GlobalData:
 
 		# interval in which a ping should be send when 
 		# no data was received/send
-		self.pingInterval = 30
+		self.pingInterval = 20
 
 		# type of this node/client
 		self.nodeType = "alert"