+++ alertClientPushNotification/lib/globalData.py
@@ -30,7 +30,7 @@ class GlobalData:
 
 		# interval in which a ping should be send when 
 		# no data was received/send
-		self.pingInterval = 30
+		self.pingInterval = 20
 
 		# type of this node/client
 		self.nodeType = "alert"