+++ alertClientPushNotification/CHANGELOG.md
@@ -3,6 +3,7 @@
 Features:
 
 * Added channel prefixing in order to reduce channel collisions.
+* Added message truncation if it is too long.
 
 
 ## 0.500