+++ alertClientPushNotification/CHANGELOG.md
@@ -4,6 +4,7 @@ Features:
 
 * Added channel prefixing in order to reduce channel collisions.
 * Added message truncation if it is too long.
+* Changed timeout setting in order to have a more stable connection recovery.
 
 
 ## 0.500